+++ R/hatmatrix.R
@@ -441,7 +441,7 @@ hatmatrix.aggr <- function(x, model, type) {
   L <- t(B) %*% W %*% B
   ## Pseudo-Inverse of L 
   L.plus <- invmat(L)
-  L.plus[is.zero(L.plus)] <- 0
+  L.plus[is_zero(L.plus)] <- 0
   ##
   ## Aggregate Hat matrix
   ##